to enable restore part of the nn
the idea is to enable restoring data for the pretrained network while
the new fc layer are still been randomly initialized. to do this
1. put all weights and bias in lenet class
2. all these weights and bias take the return value from network
building functions.
3 all network building functions return outputs, bias and weights.
3. change the save function to only save variables listed in the
function.
The software seems be able to train, resume training and run
classification correctly.  Need to use this part in the
transfer_leraning.py.