almost modular code
the code is already highly modular and fully functional. The training
and detection model  in lenet can change their outputs and training
settings according to the status of feature extraction.
The code should be further improved by defining all weights and bias in
init function of the leNet class and create new newfc and newconv
functions that take weight and bias as a vaiable instead of creating the
weights and bias internally.